feat: added UserEntityTemplate and TokenEntityTemplate in the user library
