+++ apps/backend/user-service/src/main/java/com/bytebandit/userservice/model/TokenEntity.java
@@ -1,57 +1,16 @@
 package com.bytebandit.userservice.model;
 
-import com.bytebandit.userservice.enums.TokenType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import java.sql.Timestamp;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import lib.user.model.TokenEntityTemplate;
 
 @Entity
 @Table(name = "tokens")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class TokenEntity {
-    
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    
-    @Column(name = "token_hash", nullable = false, unique = true, length = 72)
-    private String tokenHash;
-    
-    @Column(name = "is_used", nullable = false)
-    private boolean used;
-    
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, updatable = false)
-    private TokenType type;
-    
-    @Column(name = "created_at", nullable = false, updatable = false)
-    @CreationTimestamp
-    private Timestamp createdAt;
-    
-    @Column(name = "expires_at", nullable = false, updatable = false)
-    private Timestamp expiresAt;
-    
+public class TokenEntity extends TokenEntityTemplate {
+
     /**
      * Many-to-one relationship with the UserEntity. Fetches the user lazily and joins on the
      * 'user_id' column.
