+++ apps/backend/user-service/src/main/java/com/bytebandit/userservice/model/UserEntity.java
@@ -1,69 +1,16 @@
 package com.bytebandit.userservice.model;
 
 import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-import java.security.Principal;
-import java.sql.Timestamp;
-import java.util.Collection;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
+import lib.user.model.UserEntityTemplate;
 
 @Entity
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
 @Table(name = "users")
-@EntityListeners(AuditingEntityListener.class)
-public class UserEntity implements UserDetails, Principal {
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @Column(name = "email", unique = true)
-    private String email;
-
-    @Column(name = "password_hash", length = 72)
-    private String passwordHash;
-
-    @Column(name = "oauth_id")
-    private String oauthId;
-
-    @Column(name = "name")
-    private String fullName;
-
-    @Column(name = "verified", nullable = false)
-    private boolean verified = false;
-
-    @Column(name = "created_at", updatable = false)
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @Column(name = "updated_at")
-    @UpdateTimestamp
-    private Timestamp updatedAt;
+public class UserEntity extends UserEntityTemplate {
 
     /**
      * One-to-many relationship with TokenEntity. cascade = CascadeType.ALL - This means that any
@@ -75,28 +22,4 @@ public class UserEntity implements UserDetails, Principal {
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
     private Set<TokenEntity> tokens = new HashSet<>();
 
-    @Override
-    public String getName() {
-        return this.email;
-    } // take note here
-
-    @Override
-    public String getPassword() {
-        return this.passwordHash;
-    }
-
-    @Override
-    public String getUsername() {
-        return this.email;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(); // for now we don't have any roles
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return this.verified;
-    }
 }
