+++ apps/backend/user-service/src/main/java/com/bytebandit/userservice/repository/TokenRepository.java
@@ -1,10 +1,10 @@
 package com.bytebandit.userservice.repository;
 
-import com.bytebandit.userservice.enums.TokenType;
 import com.bytebandit.userservice.model.TokenEntity;
 import java.sql.Timestamp;
 import java.util.Optional;
 import java.util.UUID;
+import lib.user.enums.TokenType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
